Minimal program parsing
Basic skeletal structure of program parsing.  We ignore the contents
within the program.

Even this simple version is not correct, since a program can end with
just an `END` statement.  This can only be fixed by correctly handling
every single internal block, and then resolving the lone `END` as part
of the program.  (block and end counting is also an option, but should
give the same result.)

But this is more about testing ideas than getting a robust implementation.