Top-level construct checks; line cleanup
We now check for top-level construct checks, like DO and IF, but this
breaks horribly if there are nested constructs.  This approach will
probably need to be changed.

Continued lines are now supported, but only within the program units.
Need to create a line iterator that always does this.

A bug which was removing strings in the lowercase list comprehension was
also fixed.  (This also needs to be moved to the line iterator.)