Better two-character token support
Some two-character operators are now resolved more correctly, using
shlex's `punctuation_char` argument.  None of this is very robust,
however, and there is still probably some unexpected behaviour.

Also note that using `punctuation_chars` also adds a bunch of tokens to
words, like % and ~, which aren't used in Fortran but should probably be
dealt with at some point.  (It also adds * and / and others, but removes
them if they are explicitly declared as punctuation.)