Minimal specification statement parsing
Some basic structure for parsing specification statements has been set
up here.  There are separate functions fo the various specification
statements, but the details are not yet implemented.

We successfully exit specification on the first execution line, but we
lose this line in the next code block.  So we need some way to preserve
this final line.