Custom tokenizer replaces shlex
There was a case in shlex that was nearly unmanageable (string
delimiters in comments) which made it almost impossible to continue
using it, so I have rewritten a new tokenizer which appears to work
good enough.

There are a few cases that don't seem to work well, such as special
character combination, so this may be a step backwards, but we're in a
better position now to clean those up rather than rely on "retokenizing"
or other shlex-based trickery.