Preprocess, char context, and float parsing
Preprocessed statements are now treated like comments and bundled into a
single token.  This is mainly because C-style variables starting with _
were breaking the Fortran tokenization.

Someday, we may need to track these better to determine which #ifdef
statements to follow.

We also have improved the line continuation character context parsing,
which was failing if there was a & within the string.  We now handle
this as a special case.  This required some awkward pre-fetching of the
final token, which could cause problems someday.

Finally we have some slightly better floating point token construction.
I haven't even tested it, I only put this together because I thought it
was causing an error (which turned out to be the char context issue
above.)