Generic program unit statement test
Program unit statement checks previously looked at line[0], but this
failed for any procedure (subroutine, function) with a prefix.

We have added a static method to check this a little more rigorously
within the Unit class.  It works for most prefixes, but fails if KIND or
LEN is supplied with the declarations.

There is some commented boilerplate code in place which may fix this
someday.