Tokenizer tracks current and prior char
We now store the current and prior character during tokenization of
Fortran source.  This was necessary to sort out the string line
continuation case while still preserving the tokenization subroutine.

This also allows us to remove the char input and output arguments for
parse_string.  A similar approach should allow us to remove the
arguments from the other functions