Final removal of local char from tokenizer
The local current character (char) was removed from the last subroutine
of the tokenizer, and the objectwide self.char is now used exclusively.

I also fixed a bug related to tokenizing the kind of variables.  There
was a check to distinguish between explicit digits and named variables,
but it was apparently broken.  This new one appears to be working
better, although it's still not tested.