Partial parsing of variable declarations
Somewhat improved parsing of declarations.  Interfaces, type
definitions, and enums are now parsed as Unit blocks, mostly out of
convenience.  Subclassing may help here in the future.

There is some partial handling of actual variable declarations.  I think
we are successfully parsing the type, but the name is still unknown.