String line continuation with whitespace
Strings terminated with `&` but followed by whitespace were causing
StopIteration bugs, mostly because we were explicitly looking for `&\n`
pairs.

String tokenization is now much more accurate and explicit.  We now use
`tee` to look ahead to determine if the & denotes a line continuation or
is an explicit `&` character.

We also now rely on the standard tokenizer to handle the `&` along with
the trailing whitespace.  This will allow us to detect trailing
whitespace in the subsequent audits.

This also seems to have fixed a minor bug related to the `&` on the
subsequent line, which was incorrectly being prepended to the next part
of the string, due to a `self.prior_delim` check which was ignoring the
`&`.  We now exclude the check for this token.

Line continuations `&` and whitespace are now resolved as independent
tokens from the strings.

Notes:

- We should be using `tee` in other placese within the tokenizer, and
phase out the `prior_char` stuff.

- There is still a bug in the re-concatenation of the logical program
lines; one of the `&` is incorrectly placed between the string.
Whitespace for line continuation is also somewhat complicated and I'm
still unsure about the rules.  This still needs to be fixed.