String line continuation without matching &
String line continuation does not require a leading & on the subsequent
line, which were assuming.  We now check for the first non-whitespace
token and then appending the next token to the previous string.

There is probably opportunity to merge the two methods by checking for
whitespace and then, optionally, a `&`, which would probably be a bit
more robust.