Case check excludes comments, strings; README
Case consistency of tokens was (and probably still is) too aggressive
and was noting inconsistencies in comments and strings, for which there
is probably good reason to not worry about case.

For now, we omit these tokens from the checks, although there may be
some use for them in the future.

The README was also updated, and some sample scripts were added to the
"samples" directory.