Improved docstring support
Docstrings are now treated as tokens after whitespace and comments have
been stripped.  The docstrings are deteced by the FortLines iterator and
gathered into a docstring buffer.

Current idea is that this docstring buffer is fetched and cleared by the
various components as they are parsed.  It's a little dumb but I do not
have many options when I have not yet identified the statements.