Document class; better variable docstrings
We now track the current and previous variable when gathering
docstrings.  When the next one starts, we dump the old docstrings into
the previous variable.

This still isn't a great solution, but it works for now.

Also now gathering the "footer" material (docstrings that are floating
at the end of program units) which are used heavily in MOM6.  This
contains info that might be dumped into structures that don't normally
exist, like namespaces.

We're not yet parsing anything inside the docstrings, so there's still
opportunity here.

The header of functions is also stored, although it will later need to
be parsed into its components.

Finally, the gendoc script was updated to print most of this new
content.