Bugfix: type_decl renamed to return_type
Some changes to handling of function attributes caused me to introduce
an undefined variable (maybe a pre-existing variable, I cannot
remember.)

This patch dumps non-attributes into the return type.  It's been so long
that I'm not even sure this is what we need, but it seems to fix some
bugs.

This also seems to have some minor whitespace and code style fixes.