include preprocessor token without space
When preprocessng #include, the path does not require a space delimiter
from the directive, e.g.

    #include"some_path.h"

is a valid directive.

This patch modifies the handling to retain space delimiting for other
preprocessor directives, but splitting the path from #include.

There may be more issues with preprocessor tokenization but this seems
to fix some longstanding issues in file handling.