Better docstring anchoring
Docstrings are now gathered and stored in a list where the nth element
is comprised of all the docstrings associated with the (fully stripped
and parsed) n-1'th element.

Any subsequent docstrings on otherwise blank lines are appended to the
existing docstring.  Multilines gather up all the docstrings across the
lines into a single (LF-separated) docstring.

While this does present challenges for assiging parts of multiline
docstrings to parts of code, we can handle that case by imposing
additional structure within the docstring list.

Note that the final docstring after the final statement is not added to
self.docstring, but it is still held in self.current_doc, so can be
retrieved if needed.