Docstring restructure
Docstrings are now flagged as part of the iteration state, rather than
as a static list.

The "pre-" and "post-" docstrings are now stored in object variables
(self.current_doc, self.prior_doc) which store the docstring directly
preceding and following the current statement from the FortLines
iterator.

This information is now used to populate the Document object of various
source elements (mostly Unit types).  A "header", "docstring", and
"footer" are defined, which correspond to the docstring preceding the
object, the first docstring after the object's declaration/definition,
and the final docstring of the object.

Helper scripts will need to be modified to accommodate this new model.