Minimal command line interface started
This is not a true command line interface, this patch only lays a few
pieces required to get the command line interface working.

It sets up an argparse parser along with a single subcommand, gendoc,
which is basically hard-coded to produce MOM6 documentation out of my
dev directory.

Obviously not actually meant to be used by anyone!

ANother change is that the `tools` directory has been moved inside of
`flint` to be a proper submodule, so that the CLI can access gendoc.
This is perhaps a bit premature but is otherwise in the right direction,
I think.

More to come...