Semicolons in multiline statements
Multiline statements were always appending the next line completely,
even when the next line contained a semicolon.

This was creating unparseable statements, and also losing statements
following the semicolon.  Further chaos would occur another multiline
followed the semicolon.

This patch copypastas the semicolon handler inside the multiline block,
and terminates the staements if it contains a semicolon, moving the rest
to the buffer statement.  Any futher multilines would be resolved in the
next iteration.