Line continuation: Commas between split strings
Although our tokenizer can handle line continuations inside of strings,
it did not resolve the case of comments embedded between split strings.
This patch resolves this case by adding this exception (similar to when
& is the leading token).

While it does not fix other line continuation issues, such as splitting
of keywords, it does make the source tokenizable and the issue of
keyword splitting can be addressed in the parser.

Once again, credit to LFortran's test suite for finding this case.