Token type; case-preserving tokens
Tokens are now in a `Token` class, a subclass of str but using
case-insensitive tests (just == and hash for now; seems sufficient.)

Whitespace is also appended to the "tail" of the token (maybe renamed).
This can be used for whitespace analysis and roundrobin printing.

This new Token() based approach is much slower so I added a flag to use
the old approach, but it needs more testing.