Track leading null tokens as .head
Leading null tokens in each statement are now stored in the `head`
attribute the first token in the list of tokens for each statement.

Subsequent token heads are untracked, since they are currently used as
the tails of the prior token, but we ought to set these up at some
point.

gen_stmt was also modified to support depths for indentation, so that we
don't need to explicitly support these in the construct verbose
printing.