Line continuation inside tokens
This patch adds support for line continuation tokens inbetween semantic
tokens.  The block is triggered when `&` tokens appear at the end of a
line and the beginning of the next, and there is no whitespace before
the first & or after the second &.

When this occurs, the two tokens are concatenated into a single token,
then passed through the Tokenizer parse() function, where they either
form a new token or split back into two separate tokens.

Note that this is probably not very efficient, and a strong indication
that we probably ought to deal with line continuations before entering
FortLines, rather than trying to backtrack and re-tokenize things.

Note this breaks the `flint.use_str_as_token` flag, and probably just
needs to be phased out at this point.