Docstring added to tail of token
The docstring is now part of a token's metadata.  This improves the
roundtrip parsing to include the final docstring of a line.

It doesn't address issues with split lines or empty lines.

It also redundantly stores docstrings as both tokens and as metadata
tail tokens.  We should alter the docstring parsing to look in tail
rather than trying to iterate through the parser.  But there's quite a
bit of work ahead before that is a realistic option, starting with the
docstrings in blank lines.