Null-token lines added to tails of prior token
Blank lines and comment-only lines are now appended to the final token
of the last line to contain semantic tokens.

This does not address blank lines before the first statement, or lines
after the final statement (not sure why) but it handles most of the
normal cases.

Not finished but this is significant progress.