Rename 'parse' to 'tag'; remove old lexer block
Main change here is that the `parse` subcommand was renamed to `tag`,
although it's still relying on the verbose parser output.

A flag was also added to switch between unformatted and reformatted
output, but not yet able to be used at runtime.

Finally, the old (faster) token construction block was removed, since
other sections of the program were starting to assume the new Token()
objects.  Not terribly happy about this, but it's proven too useful to
drop.  We'll try to address performance issues later.