f90lex integration: first pass
This is a first commit of replacing the old "tokenizer" with the new
f90lex Lexer.

This new lexer gathers all whitespace and connects it to the respective
tokens.  It also preserves whitespace.  Preprocessing is integrated into
the lexer.

The downside is that the old report and docstring content had to be
stripped, since it wasn't compatible.  But it should be easier now to
gather this information and replicate the old behavior.