Better tagging with new lexer
The lexer-based output was integrated more into the tagger.  We now
print the leading whitespace of the first token, followed by all lines
of hte next statement, and finally with the everything before the final
endline of the last token.

This creates a continous reproduction of the input file with statement
tags, with two differences:

1. Missing the header whitespace (TODO)
2. Semicolons are now split, with the semicolon leading the next line.

Other little stuff too, just to keep moving:
- disable #if defined blocks (I can't yet parse complex expressions)
- some minor hack to get submodule checking to pass

Both of these issues will be returned to later.