Fix handling of interface and type declarations
This patch creates a new `Declaration` type, which parses these
multiline declarations inside of program units.

This fixes some tagging errors which were caused by the presence of such
blocks.  For example, often a `module procedure` inside of an
`interface` would incorrectly trigger a new `module` unit.

It is clear that there is much shared between Declaration and Unit, and
a lot of this probably ought to be consolidated.