Interface bugfix ; "reformat" implemented
We now explicitly require interface blocks to end with `end interface`
or `endinterface`, and no longer allow `end`.

This is consistent with the standard, although the real reason for
implementing it is that we don't yet properly handle nested blocks
inside of interfaces.

.. such as subroutine blocks terminated by `end`.

That is the real bug to fix.  But at least this lets us keep moving
ahead.

This implements a dummy version of "reformat" which really just puts
spaces between the tokens, along with a crappy indentation.