Line number support, line -> Statement
Main change here is that the lexer now returns Statements rather than
lists of tokens.  It also tracks line numbers and appends them to the
returned statements.

The end result is that report can now append line numbers.

As a side effect, most instances of `line` and `lines` are renamed to
`stmt` and `statements`.