Preprocessor #Include lexer fixes
This fixes two major and two minor problems, mostly related to include
statements.

The main problem is that queued include statements were being correctly
returned by the Lexer, but we were updating its `current_line` field,
and certain parsing operations would use a much older (non-included)
value.  This patch fixes that issue and resolves some major parsing
problems.

A second issue was that the `include` queue of the lexer was pulling
from the end via pop() which was the most recently added value, and
these were coming out in reverse (!!!).  This has been replaced with a
deque(), and we use popleft() to pull the oldest value first.

Another minor issue with preprocessing was that `#elif` was
unrecognized.  This patch handles them with `#if`, although the syntax
is still completely ignored.

A second minor issue (unrelated to #include statements) is that types
were using the subsequent token to name their blocks, which named them
`::` if that followed `type`.  We now skip past them if that's the case.
But given that types are not even units, we should not even be doing
this in Unit.parse_header.  So consider it a temporary fix.