Function result() partial parsing
This is a minor fix to the validation of the function declaration.

It is not a complete check, it only accounts for the result() final
section when validating the grammar.  It doesn't fetch the items or do
much else.

Obviously still much work to do, but it will prevent errors in function
parsing.