Statement list of lists in objects are now flat
The nested list-of-lists for the statements had advantages, but it was
very difficult to iterate over and would have required additional
iterators for convenience.

I decided it would not be worth it and is more convenient to just store
flattend lists of statements, rather than forcing people to navigate the
lists.

There are disadvantages to this approach:
- Previously, we could detect a nesting by comparing `list` to
  `Statement`.

  The most immediate impact is that I can no longer rely on this to do
  autoindentation.  I'm not too worried, since it was never really a
  good implementation anyway.  But it is a step back.

  This was always a bit dangerous anyway since `Statement` is a subclass of
  `list`.

- This increases the memory significantly.  Previously, we could just
  store a pointer to the sublist.  Now, we need to store pointers to all
  the elements of the sublist, which of course still exists in the
  subprogram/construct/etc.

I may regret this and could see myself going back to the old approach.
But for now it is the simplest approach.