Line iterator based on statements
A line iterator has been added to the project.  This take lists of
statements, each comprised of the statement's semantic tokens, and
generates the list of tokens for each line, both semantic and liminal.

This is to provide a simple interface for line-based analysis, such as
trailing whitespace.

It is not yet linked to any other modules, but should be part of the
public interface at some point.