Split tokens saved as PTokens; line length test
Tokens split across lines, such as these::

  pr&
    &i&
    &nt *, i

are now saved as PTokens.  For analysis, they test as if unsplit (e.g.
`print`) but their in-source split form is stored in its `.pp` field.
Hence, print() statements return the split form.

This was required by the line iterator, which was incorrectly merging
split tokens into single lines (and stripping their intermediate
endlines).

This was most evident in split strings, which ended up appearing as much
longer than in text.

The net reward for all this work was integration of the line length
tester, which is now functional.

Unfortunately a few linter changes also got wrapped into this PR.