Fix final comment/token without endline
The scanner assumed that all preprocessor and comment lines were
followed by an endline character.  This assumption failed if the final
line was a comment/preproc line, and it was missing the final endline
token.

This patch modifies the token formation to include, then strip, any
potential endline tokens.  The final endline is also checked before
appended as the final token.