Exclude directory support; CLI cleanup
An `--exclude` flag was added to remove subdirectories from a project
directory.

The method of passing arguments from the CLI to the tool functions was
also revised.  It is a little more rigid now, with function names now
requiring equivalents in the CLI parser, but for now it seems a bit
safer and versatile than the older approach.

None of this CLI implementation should be taken as final, it can use
some serious cleanup.