Pair token support for (/ and /)
This patch introduces tokenization for the array bound tokens (/ and /).

This was previously not possible due to conflicts with the operator
statements, e.g. `operator(/)`, `operator(//)`, etc.

We now explicitly check for operator keywords before constructing these
tokens.

We also no longer have any need for the `self.prior_char` state variable
in the Scanner, which simplifies the character iteration (although it
does not address its inherent performance issues).