Source `lines` iterator property
This patch introduces a `lines` property to `Source` to simplify the
iteration over source code lines.

The motivation is that users are going to want to iterate over lines,
will want some way to access tokens, and aren't going to want to fuss
around with statements or the `Lines` iterator.

Another advantage is that we may want to replace this with a direct IO
version at some point, rather than basing it on the stored list of
statements (which I'm already not thrilled about).

The downside here is that perhaps this is giving an impression that
`src.lines` is a list of lines, and not an iterator.