Multiline preproc #define; preproc bugfix
This patch enables support for multiline preprocessing, indicated by a
`\` as the final character.

This feature also exposed a bug where the Lexer preamble was
preprocessing the leading tokens twice.

Previously this wasn't really much of an issue since it didn't rely on
state and just preprocessed everything twice.  But now that we call
`next(self.sources)` as part of multiline reconstruction, it was
exposing the error by handling the wrong lines.

This is fixed by removing the preprocessor following `get_liminals()`,
the function which pulls in the initial "liminal" tokens before the
first statement (including any preprocessing).

A few unrelated things got shoved into this PR:

* The `debug` flag has been removed, since it wasn't doing much.

* Supported file extensions were moved to a class variable

* We now explicitly fail if a file does not contain a program unit
  (program, module, etc).  This is a precursor to anonymous main
  support.

  A `main` property was also added to the `Project`, but it's not well
  thought-out yet.