Strict module check; anonymous programs (?)
This is an unfortunate commit, but probably necessary in order to
progress on some lingering issues.

First, this replaces a broken *module-stmt* detection rule with a
stricter (and presumably accurate) correction.

This had the unexpected effect of breaking *program-stmt*, ie program
headers, which were being incorrectly tagged as modules.

Another placeholder class, `Project` was introduced to manage this case,
even though it literally adds no value.  (First mistake.)

In addition to this change, we also have some sort of perfunctory
support for anonymous main functions, i.e. files with no program unit
header (e.g. program main, module mymod, etc.)

This is "working" but incorrectly tags the first line as the header.
This is a reflection of our poor header bootstrapping, which we were
previously able to paper over.

In truth, this PR is creating further papering over of an overall poorly
designed parser.  While the Scanner and Lexer are first-rate, the Parser
itself is rather shit, with the parsing being tucked inside the various
classes.

This needs to be overhauled, probably by introducing a series of
Parser() objects which return the program components (e.g. Module,
Variable, etc) but at least this patch will tidy up a few of the sloppy
pieces, and makes the parser flaws more glaring.