Caller list generated during parsing
This patch removes the post-parsing generation of the caller list for
program units - which required an expensive sweep over the whole
codebase - and instead populates this list as the program units are
parsed.

This has a major impact on runtime performance, with speeds of large
(MOM6) codebases reduced from ~14s to ~11s.

Specifically, it reverse-populates a "caller" dict while the individuall
callee set is constructed for each program unit.

This caller dict should be seen as more of a precursor to a proper call
graph.  But for now it's sufficient.

Also I don't really like the propagation of the global "graph" across
parser functions, and may even be missing some units, but it seems to be
sufficient for now.