Compare callables against arrays
Properly declared external functions are indistinguishable from scalars,
which was causing them to be ignored in the callable function detection.

This patch creates an internal list of arrays, which are the only
possible alternative to an external callable (due to presence of `(`
following the name).  This list is now used in the callable function
tests.