Apply docstring to all comma-delimited variables
Minor change to how docstrings are assigned to variables.  We now
distribute the trailing docstring to all variables in a statement, if
they have not previously been assigned one.

This should not break the ability to assign a docstring between the
commas, nor the ability to inherit the group docstring.  It should be a
"last resort" option.

One consequence is that we now pack the statement variables into an
intermediate list (stmt_vars) which is only appended to self.variables
at the end of the statement parser.