Parenthesis count the implicit arrays
Include a more robust parenthesis counter when detecting arrays on the
RHS.  This lets me skip over any function-like content inside the size
declaration, e.g.::

    real :: x(SZI_(G), SZK_(G))

Before, it would stupidly stop on the first `)`.  That's been fixed now.

(Of course, I need to actually parse and handle that content, this is
just a quick fix to move on to some useful tasks.)